pyproject.toml
[tool.poetry]
name = "stc_test"
version = "0.1.0"
description = ""
authors = ["German Gritsenko <gerfest776@icloud.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^4.0.3"
python-decouple = "^3.6"
pre-commit = "^2.17.0"
psycopg2-binary = "^2.9.3"
djangorestframework = "^3.13.1"
drf-yasg = "^1.20.0"
mock = "^4.0.3"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
